requirements.txt
Django==5.1.7
psycopg2-binary==2.9.11
Pillow==12.1.1
django-filter==25.1
djangorestframework==3.16.1
python-dotenv==1.2.2
sqlparse==0.5.5
asgiref==3.11.1

